# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arcane Waste Management API Documentation",
        "description": "Media endpoints for media upload and management.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://arcane-2nd-version.test/api/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/media/upload": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Upload media files to S3 and optionally delete old ones",
                "operationId": "db57490fe0f8cd0f85b387cd8eab8166",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "media[]"
                                ],
                                "properties": {
                                    "media[]": {
                                        "description": "One or more media files (max 2MB each)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "remove_paths[]": {
                                        "description": "Optional array of S3 paths to delete",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Media processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Media processed successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "uploaded": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "path": {
                                                                "type": "string",
                                                                "example": "media/abc123.jpg"
                                                            },
                                                            "url": {
                                                                "type": "string",
                                                                "example": "https://bucket.s3.amazonaws.com/media/abc123.jpg"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "example.jpg"
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "example": "jpg"
                                                            },
                                                            "mime_type": {
                                                                "type": "string",
                                                                "example": "image/jpeg"
                                                            },
                                                            "size": {
                                                                "type": "integer",
                                                                "example": 204800
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "deleted": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "media/old-file.jpg"
                                                    }
                                                },
                                                "failed_deletes": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "media/fail.jpg"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - No media file provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No media file provided."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error - Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The media field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to process media",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to process media. Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}